Remove git input support; S3-only notebooks for scheduler
- Drop git clone input path, GitPython dep, and git apt package
- Read input notebooks and write executed notebooks via S3 only
- Output path: {s3_path}/runs/{name}.{runId}.ipynb
- Rename "Jupyter Gateway" to "Jupyter Container" (IOMETE naming)
- Merge fetcher.py into main.py and drop the file
