+++ jupyter-notebook-scheduler/config.yaml
@@ -8,19 +8,12 @@
 # Input Configuration
 # Defines where the notebook and dependencies are located.
 input:
-  # Type of input source. Options: "git", "s3"
-  type: "git"
+  # Type of input source. Options: "s3"
+  type: "s3"
 
   # Path to the input source.
-  # For git: The repository URL (e.g., https://github.com/org/repo.git)
   # For s3: The S3 URI (e.g., s3://my-bucket/my-project/)
-  path: "https://github.com/iomete/sample-notebooks.git"
-
-  # (Optional) Git branch or tag to checkout. Default is "main".
-  branch: "main"
-
-  # (Optional) Git token for private repositories. Pass via environment.
-  # token: "${GIT_TOKEN}"
+  path: "s3://my-bucket/my-project/"
 
 # Notebook Configuration
 notebook:
@@ -36,13 +29,13 @@ notebook:
 
 # Gateway Configuration (remote execution)
 # The notebook is executed on a remote Spark-backed kernel provisioned by the
-# IOMETE Jupyter Gateway instead of a local kernel.
+# IOMETE Jupyter Container instead of a local kernel.
 gateway:
-  # Base URL of the Jupyter Gateway / lab endpoint.
-  url: "${IOMETE_GATEWAY_URL:-https://dev.iomete.cloud/data-plane/spark-resources-1/jupyter/demo-lab}"
+  # Base URL of the Jupyter Container / lab endpoint.
+  url: "${IOMETE_JUPYTER_URL}"
 
   # Auth token for the gateway. ALWAYS provide via environment variable.
-  token: "${IOMETE_GATEWAY_TOKEN}"
+  token: "${IOMETE_JUPYTER_TOKEN}"
 
   # Auth scheme sent in the Authorization header (e.g. "token").
   auth_scheme: "token"
@@ -62,5 +55,5 @@ gateway:
 # Output Configuration
 output:
   # The S3 base path where the executed notebook is saved.
-  # The final path is constructed as: {s3_path}/{jobId}/{runId}/{filename}
-  s3_path: "s3://storage/outputs"
+  # The final path is constructed as: {s3_path}/runs/{filename}.{runId}.ipynb
+  s3_path: "s3://my-bucket/my-project/notebooks"
